Add files via upload
I am committing these files to this branch so that Viri can test something. This code is quite messy and is not intended to be merged with master. I will repost these files in a cleaner way and with more progress when I am ready for them to be merged.